requirements.txt
FastAPI
uvicorn
orjson
python-jose[cryptography]
passlib[bcrypt]
uvloop
python-dotenv
python-multipart